tests/requirements/requirements_base.txt
coverage<5
python-coveralls>2.5.0
unittest-xml-reporting==1.11.0
Pillow==5.2.0
django-treebeard>=4.3
argparse
dj-database-url
djangocms-admin-style>=1.5
django-sekizai>=0.7
django-classy-tags>=0.7.2
iptools
sphinx==1.8.5
sphinxcontrib-spelling<7.0.0 # restriction for py35 tests
pyflakes==1.1.0
pyenchant
django-app-helper
mock>=2.0.0
django-formtools
factory-boy

# Unreleased django-cms 4.0 compatible packages
https://github.com/divio/django-cms/tarball/66c70394c9e144281a0b47d93e3784d06318acf9#egg=django-cms
https://github.com/divio/djangocms-text-ckeditor/archive/support/4.0.x.zip
https://github.com/divio/djangocms-versioning/tarball/492d01c974e40b007b71e2a7e961a9019a0d5ac4#egg=djangocms-versioning
